Thermobench.jl

Julia module for working with thermobench-produced CSV files.

Installation

Run julia and type:

]develop /path/to/thermobench/julia

Reference

Thermobench.fit — Method

fit(
    time_s::Vector{Float64},
    data;
    order::Int64 = 2,
    p0 = nothing,
    tau_bounds = [(1, 60*60)],
    k_bounds = [(-120, 120)],
    T_bounds = (0, 120),
    use_cmpfit::Bool = false,
 )

Fit a thermal model to time series.

If use_cmpfit is true, use CMPFit.jl package rather than LsqFit.jl. LsqFit doesn't work well in constrained fit.

You can limit the values of fitted parameters with *_bounds parameters. Each bound is a tuple of lower and upper limit. T_bounds limits the T∞ parameter. tau_bounds and k_bounds limit the coefficients of exponential functions $k·e^{-t/τ}$. If you specify less tuples than the order of the model, the last limit will be repeated.

Example

df = read("test.csv")
f = fit(df.time_s, df.CPU_0_temp_°C)
coef(f)
Thermobench.printfit(f)

Thermobench.interpolate! — Method

interpolate!(df::AbstractDataFrame)

In-place version of interpolate.

Thermobench.interpolate — Method

interpolate!(df::AbstractDataFrame)

Replace missing values with results of linear interpolation performed against the first column (time).

julia> x = DataFrame(t=[0.0, 1, 2, 3, 1000, 1001], v=[0.0, missing, missing, missing, 1000.0, missing])
6×2 DataFrame
│ Row │ t       │ v        │
│     │ Float64 │ Float64? │
├─────┼─────────┼──────────┤
│ 1   │ 0.0     │ 0.0      │
│ 2   │ 1.0     │ missing  │
│ 3   │ 2.0     │ missing  │
│ 4   │ 3.0     │ missing  │
│ 5   │ 1000.0  │ 1000.0   │
│ 6   │ 1001.0  │ missing  │

julia> interpolate(x)
6×2 DataFrame
│ Row │ t       │ v        │
│     │ Float64 │ Float64? │
├─────┼─────────┼──────────┤
│ 1   │ 0.0     │ 0.0      │
│ 2   │ 1.0     │ 1.0      │
│ 3   │ 2.0     │ 2.0      │
│ 4   │ 3.0     │ 3.0      │
│ 5   │ 1000.0  │ 1000.0   │
│ 6   │ 1001.0  │ missing  │

Thermobench.multi_fit — Function

multi_fit(sources, columns = :CPU_0_temp_°C;
          timecol = :time_s,
          use_measurements = false,
          order::Int64 = 2,
          subtract = nothing,
          kwargs...)::MultiFit

Call fit() for all sources and report the results (coefficients etc.) in DataFrame. When use_measurements is true, report coefficients with their confidence intervals as Measurement objects.

subtract specifies the column (symbol), which is subtracted from data after interpolating its values with interpolate. This intended for subtraction of ambient temperature.

julia> multi_fit("test.csv", [:CPU_0_temp_°C :CPU_1_temp_°C])
Prefix: test.csv
2×8 DataFrame
│ Row │ name   │ column        │ mse       │ Tinf    │ k1       │ tau1    │ k2       │ tau2    │
│     │ String │ Symbol        │ Float64   │ Float64 │ Float64  │ Float64 │ Float64  │ Float64 │
├─────┼────────┼───────────────┼───────────┼─────────┼──────────┼─────────┼──────────┼─────────┤
│ 1   │        │ CPU_0_temp_°C │ 0.023865  │ 53.0003 │ -8.1627  │ 59.366  │ -13.1247 │ 317.63  │
│ 2   │        │ CPU_1_temp_°C │ 0.0208397 │ 54.0527 │ -7.17072 │ 51.1449 │ -14.3006 │ 277.687 │

Thermobench.plot_fit — Function

plot_fit(sources, columns = :CPU_0_temp_°C;
         timecol = :time_s,
         kwargs...)

Call fit for all sources and columns and produce a graph using gnuplot.

sources can be a file name (String) or a DataFrame or an array of these.

timecol is the columns with time of measurement.

Setting plotexp to true causes the individual fitted exponentials to be plotted in addition to the compete fitted function.

Other kwargs are passed to fit.

Example

plot_fit(
    ["file$i.csv" for i in 1:3],
    [:CPU_0_temp_°C, :GPU_0_temp_°C],
    order = 2
)

julia> plot_fit("test.csv", [:CPU_0_temp_°C :CPU_1_temp_°C])
rss(fit) = 16.824847736915736
rss(fit) = 14.692012083459852
LsqFit.LsqFitResult{Array{Float64,1},Array{Float64,1},Array{Float64,2},Array{Float64,1}}([54.05270193617456, -14.300579425822987, 277.6871572171704, -7.170707187492722, 51.1449427363974], [0.4814153228588438, 0.48210393166134935, -0.02827469472873645, 0.1590733915246858, -0.25639810428151577, -0.2745094012142104, -0.29552833686121005, -0.3190430919820457, 0.054645980053877, -0.1740621999878087  …  -0.10112334959897851, -0.09697970488952024, -0.09284269210126439, 0.11127121083011815, 0.11537443369089573, 0.11946698541304812, -0.0764592578095602, -0.07240014726207278, 0.1316443697488907, -0.26432565427360544], [1.0 1.0 … 1.0 -0.0; 1.0 0.9999870640586986 … 0.9999297674088051 -9.84651920490284e-6; … ; 1.0 0.07839196015996269 … 9.920656343511463e-7 -1.9227241374458278e-6; 1.0 0.07811016470837553 … 9.728568932023543e-7 -1.8881625354797446e-6], true, Float64[])

Thermobench.printfit — Method

printfit(fit; minutes = false)

Return the fitted function as Gnuplot enhanced string. Time constants (τᵢ) are sorted from smallest to largest.

Thermobench.@symarray — Macro

Construct array of symbols from arguments.

Useful for constructing column names, e.g.,

@symarray Cortex_A57_temp_°C Denver2_temp_°C

Thermobench.normalize_units! — Method

Normalize units to seconds and °C

Thermobench.read — Method

read(source; normalizeunits=true)

Read thermobech CSV file, optionally normalizing units and return it as DataFrame"

Thermobench.thermocam_correct! — Method

thermocam_correct!(df::AbstractDataFrame)

Estimate correction for thermocamera temperatures and apply it. Return the correction coefficients.

Correction is calculated from CPU_0_temp_°C and cam_cpu columns. This and the names of modified columns are currently hard coded.